% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sparklines.R
\name{renderSparkline}
\alias{renderSparkline}
\title{Widget render function for use in Shiny}
\usage{
renderSparkline(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

