% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sparklines.R
\name{sparkline}
\alias{sparkline}
\title{Sparkline htmlwidget}
\usage{
sparkline(values, chart_type = "line", config = list(), width = 100,
  height = 20)
}
\arguments{
\item{values}{A vector of values}

\item{chart_type}{The chart type: one of 'line' (default), 'bar', 'tristate', 'discrete', 'bullet', 'pie' or 'box'}

\item{config}{A named list of options as can be found on http://http://omnipotent.net/jquery.sparkline/#s-docs}
}
\description{
A sparkline htmlwidget using the nifty jQuery Sparklines library (http://omnipotent.net/jquery.sparkline)
}
\examples{
sparkline(c(1,2,-1,0,-1,-2,4,5,6), "bar", list(fillColor="white"))
}

