% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sparklines.R
\name{sparklineOutput}
\alias{sparklineOutput}
\title{Widget output function for use in Shiny}
\usage{
sparklineOutput(outputId, width = "100px", height = "20px")
}
\description{
Widget output function for use in Shiny
}

